feat: Create template module README from template
