+++ module-assets/.pre-commit-config.yaml
@@ -112,11 +112,12 @@ repos:
       entry: ci/terraform-config-inspect.sh
       language: script
       pass_filenames: false
-    - id: add_template_badges
-      name: Add badges section to README (template_module)
-      description: Add badges section to README (template_module)
-      entry: python3 ci/terraformDocBadges.py
+    - id: build_module_template_readme
+      name: Build module template README
+      description: Build module template README (template_module only)
+      entry: python3 ci/terraformDocTemplateModule.py
       language: python
+      files: "README.md"
       pass_filenames: false
 # helm lint
 - repo: https://github.com/gruntwork-io/pre-commit
