--- module-assets/.terraform-docs-config-batch.yaml
@@ -1,39 +1,0 @@
-formatter: "markdown"
-
-content: |-
-    {{- $current_repo:="" }}
-    {{- $internal_module_template_repo:="module-template" }}
-    {{- $internal_module_template_input_name:="module_template_input" }}
-    {{- range .Module.Inputs }}
-      {{- if eq .Name $internal_module_template_input_name }}
-        {{- $current_repo = $internal_module_template_repo }}
-      {{- end }}
-    {{- end }}
-
-    <!-- Update the title to match the module name and add a description -->
-    {{ if eq $internal_module_template_repo $current_repo }}# Terraform Modules template project {{ else }}# Terraform IBM Module Template {{ end }}
-    {{- if eq $internal_module_template_repo $current_repo }}
-    <!-- UPDATE BADGES:
-    1. Make sure that the badge link for the current status of the module is correct. For the status options, see https://github.ibm.com/GoldenEye/documentation/blob/master/status.md.
-    2. Update the "Build Status" badge to point to the travis pipeline for the module. Replace "module-template" in two places.
-    3. Update the "latest release" badge to point to the new module. Replace "module-template" in two places.
-    -->
-    [![Incubating (Not yet consumable)](https://img.shields.io/badge/Status-Incubating%20(Not%20yet%20consumable)-red)](https://github.ibm.com/GoldenEye/documentation/blob/master/status.md)
-    [![Build Status](https://travis.ibm.com/GoldenEye/module-template.svg?token=3Ry6sEDNvWajQPuZHgTZ&branch=master)](https://travis.ibm.com/GoldenEye/module-template)
-    [![semantic-release](https://img.shields.io/badge/%20%20%F0%9F%93%A6%F0%9F%9A%80-semantic--release-e10079.svg)](https://github.com/semantic-release/semantic-release)
-    [![pre-commit](https://img.shields.io/badge/pre--commit-enabled-brightgreen?logo=pre-commit&logoColor=white)](https://github.com/pre-commit/pre-commit)
-    [![latest release](https://shields-server.m03l6u0cqkx.eu-de.codeengine.appdomain.cloud/github/v/release/GoldenEye/module-template?logo=GitHub)](https://github.ibm.com/GoldenEye/module-template/releases/latest)
-    {{- else }}
-    <!-- UPDATE BADGE: Update the link for the badge below-->
-    [![Build Status](https://github.com/terraform-ibm-modules/terraform-ibm-module-template/actions/workflows/ci.yml/badge.svg)](https://github.com/terraform-ibm-modules/terraform-ibm-module-template/actions/workflows/ci.yml)
-    [![semantic-release](https://img.shields.io/badge/%20%20%F0%9F%93%A6%F0%9F%9A%80-semantic--release-e10079.svg)](https://github.com/semantic-release/semantic-release)
-    [![pre-commit](https://img.shields.io/badge/pre--commit-enabled-brightgreen?logo=pre-commit&logoColor=white)](https://github.com/pre-commit/pre-commit)
-    {{- end }}
-
-output:
-  file: "README.md"
-  mode: inject
-  template: |-
-    <!-- BEGIN BADGES HOOK -->
-    {{ .Content }}
-    <!-- END BADGES HOOK -->
